# apps/web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Creance",
    "version": "0.1.0",
    "description": "Parametric occupation cover funded by Displacement Bond Notes, on Hedera testnet.\n\nA worker buys cover against their occupation being displaced. Claims open when\nthe Occupation Displacement Index says the occupation is being displaced, and a\npayout also needs proof that the person lost their job. Investors fund the\npayouts and earn the premiums as coupons.\n\nTestnet only. No mainnet endpoints, no real funds.\n\nMoney: every amount is an integer string in the settlement asset smallest unit,\nwith the asset and its decimals beside it. Never parse `display`.\n\nErrors: RFC 9457 `application/problem+json`, with a `code` to switch on and a\n`retryable` boolean for an agent retry loop.\n\nPayment: the three metered operations are x402 version 2, scheme `exact`,\nnetwork `hedera:testnet`, settled through the Blocky402 testnet facilitator.\nAn unpaid call is refused with 402 and the requirements in the\n`PAYMENT-REQUIRED` header; every settlement is written to the payments topic.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://creance.co",
      "description": "Creance, Hedera testnet only."
    }
  ],
...
  "paths": {
...
    "/v1/bind": {
      "post": {
        "tags": [
          "cover"
        ],
        "operationId": "bindPolicy",
        "summary": "Bind a quoted policy",
        "description": "Registers the policy in CoverPool, writes a receipt to the payments topic\nand mints the policy NFT to the holder wallet.\n\nThe body is a quote id and nothing else: the amount, the wallet and the\nlimit all come from the quote and from the eligibility credential, so a\nrequest cannot influence its own price.\n\nAn eligibility credential is required. An agent gets one from its\nprincipal, after that person completes a World Selfie Check; the agent\nnever performs the check itself, and the NFT is always minted to the\nprincipal wallet named in the credential.\n\nThe credential is issued by POST /v1/world/verify, which forwards the\ncomplete IDKit result to World and checks it. That is the only issuer\nthis API offers a person; it is not part of this gateway because a\nSelfie Check is completed in the World App and not by an agent.\n\nPrice: the first month premium, taken from the quote. There is no static\nprice for this operation.",
        "parameters": [
          {
            "name": "PAYMENT-SIGNATURE",
            "in": "header",
            "required": false,
            "description": "x402 version 2 payment payload, network hedera:testnet, scheme exact. Required: an unpaid call is refused with 402.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "eligibilityCredential": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The bound policy.",
            "headers": {
              "PAYMENT-RESPONSE": {
                "description": "x402 version 2 settlement receipt, base64 of a JSON object with `success`, `transaction`, `network` and `payer`. The transaction id starts with the facilitator fee payer account, not with the payer, because the facilitator pays the Hedera fee; that is the id to look up on HashScan.",
                "schema": {
                  "type": "string"
                },
                "example": "eyJzdWNjZXNzIjp0cnVlLCJ0cmFuc2FjdGlvbiI6IjAuMC43MTYyNzg0QDE3ODg2MTI3MDYuNTY3NjkyNjIzIn0="
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Policy"
                }
              }
            }
          },
          "400": {
            "description": "`validation_failed` for a missing `quote_id`, `body_required` or `body_malformed` for a body that is not JSON, or `credential_ambiguous` when the header and the body carry different credentials. Unlike the other two metered operations, these are refused before the payment: the price comes from the quote, so the body is read first.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "`credential_missing`, `credential_invalid` or `credential_unknown`, which is a well-formed credential this API did not issue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "402": {
            "description": "Payment required. Price the first month premium from the quote.",
            "headers": {
              "PAYMENT-REQUIRED": {
                "description": "x402 version 2 payment requirements for this resource, base64 of the JSON `{\"x402Version\":2,\"error\":\"Payment required\",\"resource\":{...},\"accepts\":[{\"scheme\":\"exact\", \"network\":\"hedera:testnet\",\"amount\":\"10000\",\"asset\":\"0.0.10366463\",\"payTo\":\"0.0.10366450\", \"maxTimeoutSeconds\":60,\"extra\":{\"feePayer\":\"0.0.7162784\"}}]}`. The body carries the same terms in readable fields.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequired"
                }
              }
            }
          },
          "403": {
            "description": "`credential_expired`. They last thirty minutes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "`quote_not_found` or `series_not_found`. A quote that cannot be priced is refused before the 402, so the payer never signs anything.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "`already_covered`, `credential_consumed`, `quote_consumed`, `wallet_mismatch`, `group_mismatch`, `series_mismatch`, `series_not_open_for_binding` or `insufficient_capacity`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "410": {
            "description": "`quote_expired`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "`chain_write_failed`: the pool refused the write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "`hedera_not_configured`, the API has no Hedera keys and can neither publish a receipt nor mint one, or `upstream_unavailable`, the payment facilitator did not answer.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policy/{policyId}": {
      "get": {
        "tags": [
          "cover"
        ],
        "operationId": "getPolicy",
        "summary": "A policy by its id",
        "description": "Free. The response carries the cover, the dates, the NFT serial, the HCS\nreceipt sequence number and the bind transaction, and nothing that\nidentifies the person.",
        "parameters": [
          {
            "name": "policyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^pol_[0-9A-HJKMNP-TV-Z]{26}$"
            },
            "example": "pol_01K4YB9X3M8Q0RZ7T2VD6C5H9E"
          }
        ],
        "responses": {
          "200": {
            "description": "The policy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Policy"
                }
              }
            }
          },
          "400": {
            "description": "`bad_id_prefix`: that is not a policy id.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "`policy_not_found`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "eligibilityCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A short-lived eligibility credential, EdDSA over Ed25519, thirty minutes, carrying the nullifier, the occupation group, the series and the wallet. Verify it against GET /.well-known/jwks.json. An agent gets one from its principal after that person completes a World Selfie Check; the issuer is POST /v1/world/verify."
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem document. Every non-2xx response has this shape.",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "request_id",
          "retryable"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "The machine-readable reason. Switch on this, never on detail.",
            "example": "insufficient_capacity"
          },
          "request_id": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
      "PaymentRequired": {
        "type": "object",
        "description": "RFC 9457 problem document with the x402 terms repeated in readable fields.",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "retryable",
          "price",
          "x402_version",
          "scheme",
          "network",
          "pay_to",
          "facilitator"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "payment_required"
          },
          "request_id": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          },
          "x402_version": {
            "type": "integer",
            "example": 2
          },
          "scheme": {
            "type": "string",
            "example": "exact"
          },
          "network": {
            "type": "string",
            "example": "hedera:testnet"
          },
          "pay_to": {
            "type": "string",
            "description": "The Hedera account the transfer has to credit.",
            "example": "0.0.10366450"
          },
          "facilitator": {
            "type": "string",
            "description": "The x402 facilitator that settles the transfer.",
            "example": "https://api.testnet.blocky402.com"
          }
        }
      },
...
      "BindRequest": {
        "type": "object",
        "required": [
          "quote_id"
        ],
        "properties": {
          "quote_id": {
            "type": "string",
            "example": "qte_01K4YBB2R9F3M0N7X5T8W1C4Q6"
          },
          "eligibility": {
            "type": "string",
            "description": "The eligibility credential, when a client cannot set the Authorization header. Sending both is refused unless they are identical."
          }
        }
      },
      "Policy": {
        "type": "object",
        "required": [
          "policy_id",
          "series_id",
          "status",
          "limit",
          "premium"
        ],
        "properties": {
          "policy_id": {
            "type": "string",
            "example": "pol_01K4YB9X3M8Q0RZ7T2VD6C5H9E"
          },
          "series_id": {
            "type": "string",
            "example": "ODI-COMP-2026-01"
          },
          "group": {
            "type": "string",
            "enum": [
              "management_business_financial",
              "professional_related",
              "service",
              "sales_related",
              "office_admin_support",
              "farming_fishing_forestry",
              "construction_extraction",
              "installation_maintenance_repair",
              "production",
              "transportation_material_moving",
              "computer_math",
              "legal",
              "arts_design_ent_media",
              "business_financial_ops",
              "education_training_library"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "binding",
              "bound",
              "active",
              "payment_failed",
              "lapsed",
              "claims_open",
              "claimed",
              "under_review",
              "approved",
              "paid",
              "declined",
              "expired",
              "void"
            ]
          },
          "limit": {
            "$ref": "#/components/schemas/Money"
          },
          "premium": {
            "$ref": "#/components/schemas/Money"
          },
          "cover_starts": {
            "type": "string",
            "format": "date"
          },
          "cover_ends": {
            "type": "string",
            "format": "date"
          },
          "claims_payable_from": {
            "type": "string",
            "format": "date"
          },
          "next_payment_due": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "paid_through": {
            "type": "string",
            "example": "2026-09"
          },
          "holder_account": {
            "type": "string",
            "example": "0.0.10366453"
          },
          "nft": {
            "type": "object",
            "properties": {
              "token_id": {
                "type": "string",
                "nullable": true,
                "example": "0.0.10366468"
              },
              "serial": {
                "type": "integer",
                "nullable": true,
                "example": 3
              }
            }
          },
          "hcs_receipt": {
            "type": "object",
            "description": "The payments topic message that records the bind. Verify it on the mirror node at GET /topics/{topic_id}/messages?sequencenumber=eq:{sequence_number}.",
            "properties": {
              "topic_id": {
                "type": "string",
                "nullable": true,
                "example": "0.0.10366471"
              },
              "sequence_number": {
                "type": "integer",
                "nullable": true,
                "example": 41
              }
            }
          },
          "chain": {
            "type": "object",
            "properties": {
              "cover_pool": {
                "type": "string"
              },
              "bind_transaction": {
                "type": "string",
                "nullable": true
              },
              "hashscan": {
                "type": "string",
                "nullable": true
              }
            }
          },
          "premium_schedule": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string"
              },
              "href": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}